requirements.txt
openai>=2.0.0
pinecone-client>=3.0.0
python-dotenv==1.0.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-pinecone>=0.2.0
langchain-community>=0.3.0
pypdf2==3.0.1
streamlit==1.31.0

